Implemented geolocation functionality and cleaned up files
