+++ frontend/src/components/map/GoogleMap/GoogleMap.css
@@ -1,0 +1,40 @@
+.google-map-header {
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.google-map-header h2 {
+  font-size: 24px;
+  color: red;
+}
+
+.location-button {
+  padding: 10px 20px;
+  background-color: #ff3b30;
+  color: white;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+  font-size: 16px;
+  transition: background-color 0.3s ease;
+}
+
+.location-button:hover {
+  background-color: #d32f2f;
+}
+
+.google-map-container {
+  position: relative;
+  width: 80%;
+  height: 80vh;
+  margin: 0 auto;
+  margin-top: 10px;
+  background-color: white;
+  border: 5px solid red;
+  transition: all 0.3s ease;
+}
+
+.google-map-container.hovered {
+  transform: scale(1.01);
+  box-shadow: 0 0 20px rgba(255, 0, 0, 0.3);
+}
